Restructure ChannelSettings and GuildSettings
