+++ core/configuration/src/main/sqldelight/settings/database/BotSettings.sq
