+++ core/configuration/src/main/sqldelight/settings/database/ChannelSettings.sq
