+++ core/configuration/src/main/sqldelight/settings/database/GuildSettings.sq
